requirements.txt
# Google Cloud Function Requirements File
pandas>=1.3.4
numpy>=1.20.3
lxml>=4.6.3
google-cloud-storage>=2.1.0
google-cloud-bigquery>=2.32.0